fix(mobile): make the Phase 4 companion actually schedule reminders + route taps
The companion's decisive feature — exact morning/evening local
notifications — never fired: the reconciliation effect lived on
HomeScreen, which the 3-tab companion never mounts. And a notification
tap targeted AccueilTab/SymptomesTab, absent from CompanionTabParamList,
so nothing happened.

- Lift reminder reconciliation into useReminderSync(), mounted at the
  authed root (App.tsx) so it runs for BOTH the full port and the
  companion. Idempotent; same opt-ins/times. Removed the now-duplicate
  effect from HomeScreen.
- Add companionLinking: reuses the notification data URLs (home/checkin)
  but routes the morning reminder to the Matin tab and the evening
  reminder to the Soir tab.
- Document both in docs/phase4-companion.md.

pnpm typecheck (apps/mobile) → 0 errors.
